+++ ios/Classes/Video/VideoCodable.swift
@@ -8,7 +8,5 @@ import Foundation
 
 struct VideoCodable: Codable {
     var mediationName: String?
-    var userId: String?
-    var userFeature: UserFeature?
     var muteStartPlaying: Bool?
 }