+++ lib/nend_plugin.dart
@@ -3,4 +3,3 @@ export 'src/banner.dart';
 export 'src/interstitial.dart';
 export 'src/interstitial_video.dart';
 export 'src/rewarded_video.dart';
-export 'src/user_feature.dart';